Raphael Martinez: arquitectura del proyecto de Go
