+++ src/api/main.go
@@ -1,0 +1,11 @@
+package main
+
+import (
+	"github.com/gin-gonic/gin"
+)
+
+func main() {
+	router := gin.Default()
+	router.GET("/ping", helloInterface.GetNoteHandler)
+	router.Run()
+}
